requirements.txt
streamlit
langchain
chromadb
pypdf2
beautifulsoup4
requests
groq
python-dotenv
langchain_groq